pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "discussion_agents"
version = "0.1.0"
description = "Multi-agent discussion!"
authors = ["Vincent Tu <tuvincent0106@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/alckasoc/discussion-agents"
keywords = ["packaging", "poetry", "multi-agent", "LLM"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[tool.poetry.dependencies]
python = "^3.9"
cookiecutter = "^2.3.0"
openai = "^0.28.0"
tiktoken = "^0.5.1"
torch = "^2.0.1"
numpy = "^1.25.2"
langchain = "^0.0.291"
langchain-experimental = "^0.0.17"
faiss-cpu = "^1.7.4"
chromadb = "^0.4.10"
camel = {git = "https://github.com/camel-ai/camel.git", rev = "v0.1.0"}
sentence-transformers = "^2.2.2"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
python-dotenv = "^1.0.0"
black = "^23.9.1"
ruff = "^0.0.289"
ipython = "^8.15.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
pytest = "^7.4.2"

[tool.pytest.ini_options]
markers = [
    "cost: marks tests that cost funds to run (deselect with '-m \"not cost\"')",
    "slow: marks tests that are slow to run (deselect with '-m \"not slow\"')"
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.ruff]
format = "github"
select = [
    "D",  # pydocstyle
    "I",  # isort
]
target-version = "py39"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
force-wrap-aliases = true
combine-as-imports = true
lines-between-types = 1

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
exclude = ["tests/"]
